refactor: decompose ConfirmationSession from PermissionRequestLifecycleOrchestrator
